move communicate() back to six
