+++ bzt/six/__init__.py
@@ -34,3 +34,16 @@ except ImportError:
     except ImportError:
         import elementtree.ElementTree as etree
 
+
+def unicode_decode(string, errors="strict"):
+    if isinstance(string, binary_type):
+        return string.decode("utf-8", errors)
+    else:
+        return string
+
+
+def communicate(proc):  # todo: replace usage of it with sync_run()
+    out, err = proc.communicate()
+    out = unicode_decode(out, errors="ignore")
+    err = unicode_decode(err, errors="ignore")
+    return out, err
